requirements.txt

discord.py==1.7.3
openai==0.27.0
python-dotenv==0.19.2
spacy==3.2.0
aiohttp==3.8.1
nltk==3.6.7
